Exception.line is `size_t`

+++ source/kameloso/plugins/common/misc.d
@@ -237,7 +237,7 @@ final class IRCPluginSettingsException : Exception
     /// Wraps normal Exception constructors.
     this(const string message,
         const string file = __FILE__,
-        const int line = __LINE__,
+        const size_t line = __LINE__,
         Throwable nextInChain = null) pure nothrow @nogc @safe
     {
         super(message, file, line, nextInChain);
@@ -279,7 +279,7 @@ final class IRCPluginInitialisationException : Exception
     /// Wraps normal Exception constructors.
     this(const string message,
         const string file = __FILE__,
-        const int line = __LINE__,
+        const size_t line = __LINE__,
         Throwable nextInChain = null) pure nothrow @nogc @safe
     {
         super(message, file, line, nextInChain);
